Delegate handling of query templates to an extra class
